[Search] Implement SearchModel as simple ListStore
SearchModel was a custom TreeModel and had several bugs in it.
The same functionality can be achieved with a ListStore: this is
less prone to errors and should be fine for what we want.

custom models do allow for better performance in some cases, but
there's a lot we can do before we need to resort to them: e.g.
detach the model from view when adding large amounts of rows.

Signed-off-by: Jussi Kukkonen <jussi.kukkonen@intel.com>